Refactoring / tests
