+++ src/main/resources/com/cognifide/gradle/aem/test/vlt/temporaryFilter.xml
