+++ src/test/kotlin/com/cognifide/gradle/aem/test/AemTest.kt
@@ -5,10 +5,6 @@ import org.gradle.testkit.runner.GradleRunner
 import org.gradle.util.GFileUtils
 import org.junit.Rule
 import org.junit.rules.TemporaryFolder
-import org.skyscreamer.jsonassert.Customization
-import org.skyscreamer.jsonassert.JSONAssert
-import org.skyscreamer.jsonassert.JSONCompareMode
-import org.skyscreamer.jsonassert.comparator.CustomComparator
 import java.io.File
 
 abstract class AemTest {
@@ -51,22 +47,6 @@ abstract class AemTest {
 
     fun readFile(file: File): String {
         return file.bufferedReader().use { it.readText() }
-    }
-
-    fun assertJson(expected: String, actual: String) {
-        assertJsonIgnored(expected, actual, listOf())
-    }
-
-    fun assertJsonIgnored(expected: String, actual: String, ignoredFields: List<String>) {
-        val customizations = ignoredFields.map { Customization(it, { _, _ -> true }) }
-
-        assertJsonCustomized(expected, actual, customizations)
-    }
-
-    fun assertJsonCustomized(expected: String, actual: String, customizations: List<Customization>) {
-        val comparator = CustomComparator(JSONCompareMode.STRICT, *customizations.toTypedArray())
-
-        JSONAssert.assertEquals(expected, actual, comparator)
     }
 
 }
