+++ src/test/resources/com/cognifide/gradle/aem/test/vlt/cleanupDotContent/defaultClosing.xml
