+++ src/test/resources/com/cognifide/gradle/aem/test/vlt/cleanupDotContent/selfClosing.xml
